Moved formatting functions to already existing helper bean
